% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.RNNE.R
\name{predict.RNNE}
\alias{predict.RNNE}
\title{Predict from an ensemble of random neural networks}
\usage{
\method{predict}{RNNE}(models, newdata, type = "class")
}
\arguments{
\item{models}{List of models}

\item{newdata}{New data to predict outcomes}

\item{type}{Can be 'class' to predict class or 'prob' to predict probability}
}
\value{
A vector with the predictions
}
\description{
This function predicts the results of a random neural network
}
\examples{
predict.rnne(models, newdata, type = "Class")
}
\author{
Peter Xenopoulos
}
\keyword{network,}
\keyword{neural}
\keyword{prediction}
