% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.R
\name{PreProcess}
\alias{PreProcess}
\title{Preprocessing function}
\usage{
PreProcess(data, method = "zscore", ...)
}
\arguments{
\item{data}{Data frame}

\item{method}{Preprocessing method}

\item{...}{Other paramemters}
}
\value{
A transformed vector
}
\description{
This function preprocesses a vector
}
\examples{
PreProcess(data, method = "minmax")
}
\author{
Peter Xenopoulos
}
\keyword{preprocessing,}
\keyword{transform}
