% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNNE.R
\name{RNNE}
\alias{RNNE}
\title{Ensemble of Random Neural Networks}
\usage{
RNNE(formula, train, mtry = 1, total.nets = 1, verbose = FALSE,
  n.cores = 1)
}
\arguments{
\item{formula}{Formula for containing outcome and all of the predictors of interest}

\item{train}{Training dataset}

\item{mtry}{Number of predictors to randomly try}

\item{total.nets}{Number of total networks to train}

\item{verbose}{Whether or not to be verbose}

\item{n.cores}{Number of cores to use}
}
\description{
This function trains an ensemble of random neural networks using balanced bootstrapped data
}
\examples{
RNNE(formula, train, mtry = 3, total.nets = 10, verbose = TRUE, n.cores = 1)
}
\author{
Peter Xenopoulos
}
\keyword{balanced}
\keyword{bootstrap}
\keyword{ensemble,}
\keyword{network,}
\keyword{neural}
\keyword{random}
